BuffersTest: move the tests for extreme message lengths to here from `MessageTest`
It's much simpler to test the affected read helpers directly.
